+++ src/ServerApi/Infrastructures/Adnc.Infra.Helper/Core/Hash/HashGenerater.cs
@@ -1,0 +1,10 @@
+﻿namespace Adnc.Infra.Helper;
+
+public interface IHashGenerater
+{ }
+
+internal class HashGenerater : IHashGenerater
+{
+    internal HashGenerater()
+    { }
+}
