+++ src/ServerApi/Infrastructures/Adnc.Infra.Helper/Core/Security/Security.cs
@@ -1,0 +1,10 @@
+﻿namespace Adnc.Infra.Helper;
+
+public interface ISecurity
+{ }
+
+internal class Security : ISecurity
+{
+    internal Security()
+    { }
+}
