+++ src/ServerApi/Infrastructures/Adnc.Infra.Helper/Core/Security/SecurityExtentions.cs
@@ -1,0 +1,106 @@
+﻿using System.Text;
+
+namespace Adnc.Infra.Helper;
+
+/// <summary>
+/// 安全助手
+/// </summary>
+public static class SecurityExtentions
+{
+    private static readonly char[] Constant = new[]
+    {
+    '0',
+    '1',
+    '2',
+    '3',
+    '4',
+    '5',
+    '6',
+    '7',
+    '8',
+    '9',
+    'a',
+    'b',
+    'c',
+    'd',
+    'e',
+    'f',
+    'g',
+    'h',
+    'i',
+    'j',
+    'k',
+    'l',
+    'm',
+    'n',
+    'o',
+    'p',
+    'q',
+    'r',
+    's',
+    't',
+    'u',
+    'v',
+    'w',
+    'x',
+    'y',
+    'z'
+};
+
+    private static readonly char[] ConstantNumber = new[]
+    {
+    '0',
+    '1',
+    '2',
+    '3',
+    '4',
+    '5',
+    '6',
+    '7',
+    '8',
+    '9'
+};
+
+    public static readonly Random Random = new();
+
+    /// <summary>
+    /// 生成随机验证码
+    /// </summary>
+    /// <param name="length">验证码长度</param>
+    /// <param name="isNumberOnly">验证码是否是纯数字</param>
+    /// <returns></returns>
+    public static string GenerateRandomCode(this ISecurity _, int length, bool isNumberOnly = false)
+    {
+        char[] array;
+        if (isNumberOnly)
+        {
+            array = ConstantNumber;
+        }
+        else
+        {
+            array = Constant;
+        }
+        var stringBuilder = new StringBuilder(length);
+        for (var i = 0; i < length; i++)
+        {
+            //stringBuilder.Append(array[Random.Next(array.Length)]);
+            stringBuilder.Append(array[Random.StrictNext(array.Length)]);
+        }
+        return stringBuilder.ToString();
+    }
+
+    /// <summary>
+    /// get MD5 hashed string
+    /// </summary>
+    /// <param name="sourceString">原字符串</param>
+    /// <param name="isLower">加密后的字符串是否为小写</param>
+    /// <returns>加密后字符串</returns>
+    public static string MD5(this ISecurity _, string sourceString, bool isLower = false)
+    {
+        if (string.IsNullOrEmpty(sourceString))
+        {
+            return "";
+        }
+        return InfraHelper.Hash.GetHashedString(HashType.MD5, sourceString, isLower);
+    }
+}
