+++ src/ServerApi/Infrastructures/Adnc.Infra.Helper/InfraHelper.cs
@@ -1,0 +1,11 @@
+﻿namespace Adnc.Infra.Helper;
+
+public class InfraHelper
+{
+    private InfraHelper()
+    { }
+
+    public static ISecurity Security => new Security();
+
+    public static IHashGenerater Hash => new HashGenerater();
+}
